docs(example): add channel example
This example utilizes a tokio::sync::mpsc::channel for communicating
events. This has the benefit of vastly simplifying code: no application
logic has to exist inside of the `async_debounce_watch` function.
Instead, the debouncer is kept in scope because the function returns it
as well as the receiver used to retrieve events.