feat: add debounce functions that use a tokio::sync::mpsc::channel
Following inspiration from the channel example, functions have been
implemented that return an async channel as well as the debouncer. This
simplifies many scenarios and allows for the debouncer to be used
without much setup.

There doesn't appear to be a simple and straightforward way to implement
this for other types of channels. For now the regular tokio channels
will do. Consumers are still able to use alternate channel
implementations; they'll just have to add a small bit of boilerplate. In
the future some sort of generic channel could be utilized.