fix(cli): support commands that exit
Previously, if the invoked command exited the application would panic.
This is because the application loop would poll the task's future in the
`select!` macro, but the application had exited and there was no future
to poll.

This change makes it so that a new task is spawned that does nothing
except wait for the CancellationToken signal. This means that in
subsequent loop iterations there will be no child command running, but
the moment that file events are received the task is terminated and the
child command is executed again.